šimek/pyproject.toml
[project]
name = "simek-bot"
version = "0.1.0"
description = "Šimek bot."
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "aiohttp>=3.11.11,<4.0.0",
    "requests>=2.32.5,<3.0.0",
    "disnake>=2.10.1,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "ufal-morphodita==1.11.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=1.2.0,<1.3.0",
    "pytest-mock>=3.12.0,<4.0.0",
    "ruff>=0.13.2,<1.0.0",
    "pytest-mypy>=0.10.3,<1.1.0",
    "pytest-ruff>=0.3.1,<0.6.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = ["F403", "F405", "E722", "F841"]

[tool.mypy]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "disnake.*"
follow_imports = "skip"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--ruff"]
#addopts = ["-v", "--tb=short", "--ruff", "--mypy"]
